Prevents a double enqueue
It was possible for a user to enqueue, then wait five minutes, then
enqueue again with a different BattleTag. This was possible because the
queue itself holds the bytes of their user id, guild id, and battle tag,
so detecting the duplicate wasn't possible. But now, if the user
specifies their battle tag on the command line, we check for an existing
cached battle tag, and if found, we check if that combination is already
in the queue, thus preventing the loop hole.